Add better output

+++ MI.Server.BusinessLogic/DTO/SubmissionDto.cs
@@ -16,7 +16,7 @@ namespace MI.Server.BusinessLogic.DTO
 
         public int NumberOfUploads { get; set; }
 
-        public int RunTime { get; set; }
+        public double RunTime { get; set; }
 
         public UserDto User { get; set; }
